Add main structure for Codingame bot Chrome extension
This commit involves the implementation of the main structure for a Codingame Bot Chrome extension. This extension automatically solves Codingame exercises. The structure includes JavaScript for API communication, DOM manipulation, and interaction with the content of the webpage, manifest.json, .gitignore, package.json and tsconfig.json for package setup, and HTML and CSS for the extension's UI.